Added LSTM example incomplete
Added LSTM example duplicated from GAN although still needs reworking to
become fully functional LSTM. This will replace the inbuilt LSTM in
Nemesyst as the architecture used in the GAN example is superior and it
had a lot more foresight tha the original LSTM implementation, in
particular individual batches and epochs. and generalising thngs better.
This also takes out the neural network from nemesyst core code so that
it can be slimmed.