Update lstm.py
Updated lstm.py with prediction and corrected a mistake with tester.
Predictions currentley output in current directory but I will add
database saving soon into these two examples. So whats next is Saving
the models to the database using gridfs, and saving model output in
gridfs, then linking the model, the model metadata, model output, and
trained on data together to create a whole history of a model such that
it is easy to trace things back and reuse them in future. This should
probably also be followed with examples of reopening models from the
database in keras.