Updated nemesyst and mongo_handler
updated both to match keys, and fixed an issue with connections, and
added ability to change unauth database port rather than being all hard
coded but localhost is still hard coded for init since we never want it
exposed to the world. Also updated naming convention to be full port not
part form.