Added nemesyst generator import and tests
tests arent 100% there they do not check that the data is retrievable
again but it is from manual testing. Now user scripts must yield what
they want to import into the database. this way it is easy for them to
make their scripts loop and import everything. The next issue will be
gridfs.