Updated nemesyst parallelisation
nemesyst now properly deep copies arguments and creates new databases
for each operation of scripts such that we can now add parallelisation
so each script can be run independantly. Then we will need to pass in
counters to the scripts so that they can know which one they are and how
many counters there will be total.