Fixed pymongo duplicate key -> operation fail
so pymongo 3.9.0-1 has landed but the error for a user already existing
has changed to be operation failure BUT this is already used for normal
operation failures so we have to split this error into the two again
ourselves since who care if the user already exists? just continue with
our lives.