Updated config options for MNIST + nemesyst
updated the options with better descriptions and added better spacing.
Also adding new option triggers. The only caveat is I will need to
rework password as its pretty annoying as it currentley stands. I think
it would be best to have it be able to be set from plaintext file
despite the security risk pureley for experimentation so people dont get
confused why they are being asked for 60 different passwords.