Updated mnist learner to be able to fetch data
This serves as a good example of fetching data from the database, but it
also made me realise maybe a two part call to get a cursor and use it
should be abstracted into a single call instead. As its not clear to
user whats going on that infact there is a referance being held in the
backend.