Added validation generator to mnist example
This needs to now use pipelines to select only data that is desired but
its almost done! Now it will be necessary to use pipelines to split into
train and test data. This will also require that I get the total number
of steps automagically but it works as is.