Added dl-sequence-length arg
This argument is another utility argument for specifying the sequence
length to any sub scripts so they can dynamically adjust pipelines and
data processing, and neural networks to handle a specific sequence
length. This is not strictly needed but is a handy utility to have. In
future I may consider having sub config files being passable so that
they dont need to set args through nemesyst.